rename packages in clusternet controller manager
