+++ pkg/controllermanager/deployer/localizer/patch.go
