+++ pkg/controllermanager/deployer/localizer/patch_test.go
